chore: fixing imports and moving cached_lookup to stark-backend
