+++ stark-backend/tests/cached_lookup/mod.rs
@@ -11,7 +11,7 @@ use p3_matrix::dense::RowMajorMatrix;
 use p3_util::log2_ceil_usize;
 
 use crate::config;
-use crate::interaction::dummy_interaction_air::DummyInteractionAir;
+use afs_test_utils::interaction::dummy_interaction_air::DummyInteractionAir;
 
 type Val = BabyBear;
 
@@ -107,4 +107,83 @@ pub fn prove_and_verify_indexless_lookups(
         proof,
         &pis,
     )
+}
+
+/// tests for cached_lookup
+#[test]
+fn test_interaction_cached_trace_happy_path() {
+    // count fields
+    //   0    1 1
+    //   7    4 2
+    //   3    5 1
+    // 546  889 4
+    let sender = vec![
+        (0, vec![1, 1]),
+        (7, vec![4, 2]),
+        (3, vec![5, 1]),
+        (546, vec![889, 4]),
+    ];
+
+    // count fields
+    //   1    5 1
+    //   3    4 2
+    //   4    4 2
+    //   2    5 1
+    //   0  123 3
+    // 545  889 4
+    //   1  889 4
+    //   0  456 5
+    let receiver = vec![
+        (1, vec![5, 1]),
+        (3, vec![4, 2]),
+        (4, vec![4, 2]),
+        (2, vec![5, 1]),
+        (0, vec![123, 3]),
+        (545, vec![889, 4]),
+        (1, vec![889, 4]),
+        (0, vec![456, 5]),
+    ];
+
+    prove_and_verify_indexless_lookups(sender, receiver).expect("Verification failed");
+}
+
+#[test]
+fn test_interaction_cached_trace_neg() {
+    // count fields
+    //   0    1 1
+    //   7    4 2
+    //   3    5 1
+    // 546  889 4
+    let sender = vec![
+        (0, vec![1, 1]),
+        (7, vec![4, 2]),
+        (3, vec![5, 1]),
+        (546, vec![889, 4]),
+    ];
+
+    // field [889, 4] has count 545 != 546 in sender
+    // count fields
+    //   1    5 1
+    //   3    4 2
+    //   4    4 2
+    //   2    5 1
+    //   0  123 3
+    // 545  889 4
+    //   1  889 10
+    //   0  456 5
+    let receiver = vec![
+        (1, vec![5, 1]),
+        (3, vec![4, 2]),
+        (4, vec![4, 2]),
+        (2, vec![5, 1]),
+        (0, vec![123, 3]),
+        (545, vec![889, 4]),
+        (1, vec![889, 10]),
+        (0, vec![456, 5]),
+    ];
+
+    assert_eq!(
+        prove_and_verify_indexless_lookups(sender, receiver),
+        Err(VerificationError::NonZeroCumulativeSum)
+    );
 }
